refactor(extension-file): minor update
- move `uploadFile` to a separate file.
- add generic type so that we can support different types of node in the future.
